+++ android/sdl_android/src/main/java/com/smartdevicelink/managers/SdlManager.java
@@ -35,8 +35,6 @@ package com.smartdevicelink.managers;
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.os.Build;
-import android.os.Handler;
-import android.os.Looper;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.util.Log;
@@ -45,7 +43,6 @@ import com.smartdevicelink.exception.SdlException;
 import com.smartdevicelink.managers.audio.AudioStreamManager;
 import com.smartdevicelink.managers.file.FileManager;
 import com.smartdevicelink.managers.file.filetypes.SdlArtwork;
-import com.smartdevicelink.managers.lifecycle.LifecycleConfigurationUpdate;
 import com.smartdevicelink.managers.lockscreen.LockScreenConfig;
 import com.smartdevicelink.managers.lockscreen.LockScreenManager;
 import com.smartdevicelink.managers.permission.PermissionManager;
@@ -55,7 +52,6 @@ import com.smartdevicelink.protocol.enums.FunctionID;
 import com.smartdevicelink.protocol.enums.SessionType;
 import com.smartdevicelink.proxy.RPCMessage;
 import com.smartdevicelink.proxy.RPCRequest;
-import com.smartdevicelink.proxy.RPCResponse;
 import com.smartdevicelink.proxy.SdlProxyBase;
 import com.smartdevicelink.proxy.SystemCapabilityManager;
 import com.smartdevicelink.proxy.callbacks.OnServiceEnded;
@@ -65,7 +61,6 @@ import com.smartdevicelink.proxy.interfaces.ISdl;
 import com.smartdevicelink.proxy.interfaces.ISdlServiceListener;
 import com.smartdevicelink.proxy.interfaces.IVideoStreamListener;
 import com.smartdevicelink.proxy.interfaces.OnSystemCapabilityListener;
-import com.smartdevicelink.proxy.rpc.ChangeRegistration;
 import com.smartdevicelink.proxy.rpc.OnHMIStatus;
 import com.smartdevicelink.proxy.rpc.RegisterAppInterfaceResponse;
 import com.smartdevicelink.proxy.rpc.SdlMsgVersion;
@@ -74,14 +69,12 @@ import com.smartdevicelink.proxy.rpc.TTSChunk;
 import com.smartdevicelink.proxy.rpc.TemplateColorScheme;
 import com.smartdevicelink.proxy.rpc.enums.AppHMIType;
 import com.smartdevicelink.proxy.rpc.enums.Language;
-import com.smartdevicelink.proxy.rpc.enums.Result;
 import com.smartdevicelink.proxy.rpc.enums.SdlDisconnectedReason;
 import com.smartdevicelink.proxy.rpc.enums.SystemCapabilityType;
 import com.smartdevicelink.proxy.rpc.listeners.OnMultipleRequestListener;
 import com.smartdevicelink.proxy.rpc.listeners.OnRPCListener;
 import com.smartdevicelink.proxy.rpc.listeners.OnRPCNotificationListener;
 import com.smartdevicelink.proxy.rpc.listeners.OnRPCRequestListener;
-import com.smartdevicelink.proxy.rpc.listeners.OnRPCResponseListener;
 import com.smartdevicelink.security.SdlSecurityBase;
 import com.smartdevicelink.streaming.audio.AudioStreamingCodec;
 import com.smartdevicelink.streaming.audio.AudioStreamingParams;
@@ -92,8 +85,6 @@ import com.smartdevicelink.transport.enums.TransportType;
 import com.smartdevicelink.transport.utl.TransportRecord;
 import com.smartdevicelink.util.DebugTool;
 import com.smartdevicelink.util.Version;
-
-import org.json.JSONException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -114,8 +105,6 @@ import java.util.Vector;
  */
 public class SdlManager extends BaseSdlManager{
 	private static final String TAG = "SdlManager";
-	private static final int MAX_RETRY = 3;
-	private int changeRegistrationRetry = 0;
 	private SdlProxyBase proxy;
 	private SdlArtwork appIcon;
 	private Context context;
@@ -139,6 +128,7 @@ public class SdlManager extends BaseSdlManager{
 			DebugTool.logInfo("Proxy is connected. Now initializing.");
 			changeRegistrationRetry = 0;
 			checkLifecycleConfiguration();
+			initialize();
 		}
 
 		@Override
@@ -241,74 +231,7 @@ public class SdlManager extends BaseSdlManager{
 
 	@Override
 	protected void checkLifecycleConfiguration(){
-
-		Language actualLanguage =  proxy.getRegisterAppInterfaceResponse().getLanguage();
-		LifecycleConfigurationUpdate lcu = managerListener.managerShouldUpdateLifecycle(actualLanguage);
-
-		if (lcu != null){
-
-			// go through and change sdlManager properties that were changed via the LCU update
-			hmiLanguage = actualLanguage;
-
-			if (lcu.getAppName() != null){
-				appName = lcu.getAppName();
-			}
-
-			if (lcu.getShortAppName() != null){
-				shortAppName = lcu.getShortAppName();
-			}
-
-			if (lcu.getTtsName() != null){
-				ttsChunks = lcu.getTtsName();
-			}
-
-			if (lcu.getVoiceRecognitionCommandNames() != null){
-				vrSynonyms = lcu.getVoiceRecognitionCommandNames();
-			}
-
-			ChangeRegistration changeRegistration = new ChangeRegistration(actualLanguage,actualLanguage);
-			changeRegistration.setAppName(lcu.getAppName());
-			changeRegistration.setNgnMediaScreenAppName(lcu.getShortAppName());
-			changeRegistration.setTtsName(lcu.getTtsName());
-			changeRegistration.setVrSynonyms(lcu.getVoiceRecognitionCommandNames());
-			changeRegistration.setOnRPCResponseListener(new OnRPCResponseListener() {
-				@Override
-				public void onResponse(int correlationId, RPCResponse response) {
-					try {
-						DebugTool.logInfo(response.serializeJSON().toString());
-					} catch (JSONException e) {
-						e.printStackTrace();
-					}
-				}
-
-				@Override
-				public void onError(int correlationId, Result resultCode, String info){
-					DebugTool.logError( "Change Registration onError: "+ resultCode+ " | Info: "+ info );
-					changeRegistrationRetry++;
-					if (changeRegistrationRetry < MAX_RETRY) {
-						final Handler handler = new Handler(Looper.getMainLooper());
-						handler.postDelayed(new Runnable() {
-							@Override
-							public void run() {
-								checkLifecycleConfiguration();
-								DebugTool.logInfo("Retry Change Registration Count: "+ changeRegistrationRetry);
-							}
-						}, 3000);
-					}
-				}
-			});
-			try {
-				proxy.sendRPC(changeRegistration);
-			} catch (SdlException e) {
-				e.printStackTrace();
-			}
-		}
-
-		// we don't really care about that response, initialize the managers
-		// This will get called regardless of the retry count, so only send it once.
-		if (changeRegistrationRetry == 0) {
-			initialize();
-		}
+		this.checkLifecycleConfigurationPrivate(managerListener);
 	}
 
 	@Override
